# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Company news component",
    "description": "Standalone news backend. Resolves every catalogue ticker to its exchange and to the full set of venues it trades on, then keeps a deduplicated news feed per company.",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/v1/runs/{id}": {
      "get": {
        "tags": [
          "ops"
        ],
        "description": "One run, with its provider breakdown.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/runs/{id}/companies": {
      "get": {
        "tags": [
          "ops"
        ],
        "description": "Per-company outcomes for a run. Filter by outcome to separate 'no_news' from 'refused'.",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    }
  },
...
}